Fix a ton of issues
First of all, fix a bunch of edge cases I hadn't thought of. Some of them:

  * Anonymous classes
  * Test functions
  * Variable names in lambdas

Instead of checking that a variable name meets this specific pattern, check
that the name matches the bare requirements. For example, if a variable name
has starts with an uppercase letter, has a lowercase letter, and does not have
an underscore it is probably an UpperCamel case.

I have tested this against Teku and I think the results look good.