Add MathTargetType check (#7)
This will identify calls to Math functions (min, max, addExact, etc) where the
input types are different than the result type. For example, if a long variable
is set to the minimum of two integers. In this situation the result could just
be an int. It is also possible that there was a mistake; for example 1 vs 1L.
The errorprone IntLongMath check is great about finding those, but it is
limited to what it knows.